Move asset manager logic to specific folder
