+++ server/api/socket/asset_manager/common.py
@@ -1,0 +1,17 @@
+from typing_extensions import TypedDict
+
+from utils import FILE_DIR
+
+
+class UploadData(TypedDict):
+    uuid: str
+    name: str
+    directory: int
+    slice: int
+    totalSlices: int
+    data: bytes
+
+
+ASSETS_DIR = FILE_DIR / "static" / "assets"
+if not ASSETS_DIR.exists():
+    ASSETS_DIR.mkdir()
